% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyNaiveBayes-package.R
\docType{package}
\name{FuzzyNaiveBayes-package}
\alias{FuzzyNaiveBayes}
\alias{FuzzyNaiveBayes-package}
\title{FuzzyNaiveBayes: Fuzzy Classifiers}
\description{
Fuzzy Classifiers for data.
}
\author{
\strong{Maintainer}: Jodavid Ferreira \email{jodavid@protonmail.com} (\href{https://orcid.org/0000-0002-2131-6464}{ORCID})

Other contributors:
\itemize{
  \item Ronei Moraes \email{ronei@de.ufpb.br} [contributor]
}

}
\keyword{internal}
