% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GamWeightData.R
\docType{data}
\name{GamWeightData}
\alias{GamWeightData}
\title{Gamma Weighted Data}
\format{
A dataset with 600 rows and 4 variables with 1 label.
}
\usage{
GamWeightData
}
\description{
A dataset simulated containing training data from a Gamma Distribution
}
\keyword{datasets}
