% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaiveBayesClassifiers-package.R
\docType{package}
\name{NaiveBayesClassifiers-package}
\alias{NaiveBayesClassifiers}
\alias{NaiveBayesClassifiers-package}
\title{NaiveBayesClassifiers: Fuzzy e Non-Fuzzy Naive Bayes Classifiers}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Fuzzy Classifiers for data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Jodavid/NaiveBayesClassifiers}
  \item Report bugs at \url{https://github.com/Jodavid/NaiveBayesClassifiers/issues}
}

}
\author{
\strong{Maintainer}: Jodavid Ferreira \email{jodavid@protonmail.com} (\href{https://orcid.org/0000-0002-2131-6464}{ORCID})

Other contributors:
\itemize{
  \item Ronei Moraes \email{ronei@de.ufpb.br} [contributor]
}

}
\keyword{internal}
