% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatedData.R
\docType{data}
\name{SimulatedData}
\alias{SimulatedData}
\title{Simulated Data}
\format{
A dataset with 600 rows and 4 variables with 1 label.
}
\usage{
SimulatedData
}
\description{
A dataset containing training data from Gammma Distribuition
}
\keyword{datasets}
