% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirtualRealityData.R
\docType{data}
\name{VirtualRealityData}
\alias{VirtualRealityData}
\title{Virtual Reality Simulator Data}
\format{
A dataset with 600 rows and 4 variables with 1 label.
}
\usage{
VirtualRealityData
}
\description{
A dataset containing training data from a virtual reality simulator
}
\keyword{datasets}
