% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyClass-package.R
\docType{package}
\name{FuzzyClass-package}
\alias{FuzzyClass}
\alias{FuzzyClass-package}
\title{FuzzyClass: Fuzzy and Non-Fuzzy Classifiers}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides classifiers that can be used for discrete variables and for continuous variables based on the idea of Naive Bayes and Fuzzy Naive Bayes considering some statistical distributions of articles published in the literature developed in the LabTEVE and LEAPIG research laboratories. Among the proposed classification methods is a with the Gamma distribution, proposed by Moraes, Soares and Machado (2018) <doi:10.1142/9789813273238_0088>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Jodavid/FuzzyClass}
  \item Report bugs at \url{https://github.com/Jodavid/FuzzyClass/issues}
}

}
\author{
\strong{Maintainer}: Jodavid Ferreira \email{jodavid@protonmail.com} (\href{https://orcid.org/0000-0002-2131-6464}{ORCID})

Other contributors:
\itemize{
  \item Ronei Moraes \email{ronei@de.ufpb.br} (\href{https://orcid.org/0000-0001-8436-8950}{ORCID}) [contributor]
  \item Arthur Ricardo \email{arthur.ricardo17@gmail.com} [contributor]
}

}
\keyword{internal}
