% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyClass-package.R
\docType{package}
\name{FuzzyClass-package}
\alias{FuzzyClass}
\alias{FuzzyClass-package}
\title{FuzzyClass: Fuzzy and Non-Fuzzy Classifiers}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

It provides classifiers which can be used for discrete variables and for continuous variables based on the Naive Bayes and Fuzzy Naive Bayes hypothesis. Those methods were developed by researchers belong to the 'Laboratory of Technologies for Virtual Teaching and Statistics (LabTEVE)' and 'Laboratory of Applied Statistics to Image Processing and Geoprocessing (LEAPIG)' at 'Federal University of Paraiba, Brazil'. They considered some statistical distributions and their papers were published in the scientific literature, as for instance, the Gaussian classifier using fuzzy parameters, proposed by 'Moraes, Ferreira and Machado' (2021) \doi{10.1007/s40815-020-00936-4}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leapigufpb/FuzzyClass}
  \item Report bugs at \url{https://github.com/leapigufpb/FuzzyClass/issues}
}

}
\author{
\strong{Maintainer}: Jodavid Ferreira \email{jodavid@protonmail.com} (\href{https://orcid.org/0000-0002-2131-6464}{ORCID})

Other contributors:
\itemize{
  \item Ronei Moraes \email{ronei@de.ufpb.br} (\href{https://orcid.org/0000-0001-8436-8950}{ORCID}) [contributor]
}

}
\keyword{internal}
