% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyNaiveBayes-package.R
\docType{package}
\name{NaiveBayesClassifiers-package}
\alias{NaiveBayesClassifiers}
\alias{NaiveBayesClassifiers-package}
\title{NaiveBayesClassifiers: Fuzzy e non-fuzzy Naive Bayes Classifiers}
\description{
Fuzzy Classifiers for data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Jodavid/FuzzyNaiveBayes}
  \item Report bugs at \url{https://github.com/Jodavid/FuzzyNaiveBayes/issues}
}

}
\author{
\strong{Maintainer}: Jodavid Ferreira \email{jodavid@protonmail.com} (\href{https://orcid.org/0000-0002-2131-6464}{ORCID})

Other contributors:
\itemize{
  \item Ronei Moraes \email{ronei@de.ufpb.br} [contributor]
}

}
\keyword{internal}
